requirements.txt
requests==2.25.1
requests-oauthlib==1.3.0
paho-mqtt==1.5.1
